requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
pydantic-settings
python-dotenv
python-multipart
httpx
supabase
pyjwt
pandas
sqlalchemy
psycopg2-binary
statsmodels
matplotlib
pytest
pytest-asyncio
